requirements.txt
requests
Flask
Flask-WTF
Flask-Bootstrap4
WTForms
